refactor(storage): Widen SearchCapabilities beyond which fields are required
The type sat in its own file describing only which query fields a reader accepts
as empty, which is narrower than what backends actually diverge on.

It now lives beside the Reader interface it belongs to, and its doc says what the
type is for: how faithfully a reader serves a search — fields that may be
omitted, fields honored exactly rather than approximated, and combinations a
backend cannot serve. Two divergences that exist today are named as the fields to
expect, with references: SearchDepth is an exact limit on some backends and a
hint on others, which jaeger.api_v3 already warns about, and Cassandra cannot
intersect a duration filter with tags because durations come from a separate
duration_index whose partition key does not compose (ADR-001, #1047, and #1015
on what a duration filter means at all).

The tripwire test moves into reader_test.go with the type.
